+++ tests/test_impl/__init__.py
@@ -2,6 +2,7 @@
 
 from .test_claude import *
 from .test_codex import *
+from .test_codex_reconnect import *
 from .test_grok import *
 from .test_omp import *
 from .test_opencode import *